revive simplifier rule (T)(c?x:y) --> c?(T)x:(T)y
This rule was considered in the past, but removed again as this
simplification may yield a more expensive formula if the complexity of 'T'
is higher than that of the if expression (say 'double' vs.  'int').

Thus, we only enable when x and y are constant.  This enables re-writing,
for instance, (long)c?1:0 to c?1L:0L.